add history DB
