+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "pytest-fly"
-description = "pytest observer"
-version = "0.0.25"
+description = "pytest runner and observer"
+version = "0.0.26"
 readme = "README.md"
 requires-python = ">=3.12"
 authors = [
