+++ src/pytest_fly/app/model/__init__.py
@@ -1,1 +1,1 @@
-from .pytest_process import PytestResult, PytestProcessState, PytestStatus, PytestKey, exit_code_to_string, key_from_pytest_status
+from .pytest_process import PytestKey, exit_code_to_string, key_from_pytest_status
