+++ src/pytest_fly/app/view/about.py
@@ -4,7 +4,7 @@ from PySide6.QtCore import QObject, QThread, Signal
 from PySide6.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout
 from .gui_util import PlainTextWidget
 from ..project_info import get_project_info
-from ..platform_info import get_platform_info, get_performance_core_count
+from pytest_fly.common.platform_info import get_platform_info, get_performance_core_count
 
 
 class AboutDataWorker(QObject):
