+++ src/pytest_fly/plugin/__init__.py
@@ -1,0 +1,1 @@
+from .pytest_fly_plugin import pytest_addoption, pytest_runtest_logreport, pytest_sessionfinish, pytest_sessionstart
