+++ src/pytest_fly/plugin/pytest_fly_plugin.py
@@ -1,6 +1,6 @@
 import pytest
 from _pytest.reports import BaseReport
-from .db import write_report, write_start, write_finish
+from ..db import write_report, write_start, write_finish
 
 
 def pytest_addoption(parser):
