+++ tests/test_pytest_runner.py
@@ -4,8 +4,8 @@ import time
 from PySide6.QtCore import QThread
 from pytest import ExitCode
 from pytest_fly.app.controller import PytestRunnerWorker
-from pytest_fly.app.model import PytestProcessState
-from pytest_fly.app.platform_info import get_performance_core_count
+from pytest_fly.common.classes import PytestProcessState
+from pytest_fly.common.platform_info import get_performance_core_count
 
 
 def test_pytest_runner(app):
